pyproject.toml
# see https://flit.pypa.io/en/latest/pyproject_toml.html
[build-system]
#requires = ["flit_core >=3.2,<4"]
#build-backend = "flit_core.buildapi"
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pyMetaModel"
authors = [
  {name = "Wolfgang Fahl", email = "wf@bitplan.com"}
]
maintainers = [
  { name = "Wolfgang Fahl", email = "wf@bitplan.com" },
]
readme = "README.md"
# flit_core.config.ConfigError: license field should be <class 'dict'>, not <class 'str'>
license = { file="LICENSE" }
dependencies = [
		# https://pypi.org/project/py-sidif/
        'py-sidif~=0.4.0',
        # https://pypi.org/project/py-3rdparty-mediawiki/
		'py-3rdparty-mediawiki~=0.7.0',
		# https://pypi.org/project/mwparserfromhell/
		'mwparserfromhell~=0.6.4'
     ]

requires-python = ">=3.7"
classifiers=[
    "Development Status :: 4 - Beta",
    "Environment :: Web Environment",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Operating System :: OS Independent",
    "Topic :: Software Development :: Meta Model",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License"
    ]
dynamic = ["version", "description"]
[tool.hatch.version]
path = "meta/__init__.py"

[project.urls]
Home = "https://github.com/WolfgangFahl/pyMetaModel"
Documentation = "https://wiki.bitplan.com/index.php/PyMetaModel"
Source = "https://github.com/WolfgangFahl/pyMetaModel"

[project.optional-dependencies]
test = [
  "green",
]

[tool.hatch.build.targets.wheel]
packages = [
  "meta",
]

